src/test/java/com/rightfindpro/become/mapper/PageDtoMapperTest.java
package com.rightfindpro.become.mapper;

import com.rightfindpro.become.dto.PageDto;
import org.junit.jupiter.api.Test;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;

import java.util.Arrays;
import java.util.Iterator;
import java.util.List;
import java.util.function.Function;

import static org.junit.jupiter.api.Assertions.*;

class PageDtoMapperTest {


    @Test
    void toPageResponse() {
        PageDtoMapper pageDtoMapper = new PageDtoMapper();
        PageDto pageDto = pageDtoMapper.toPageResponse(new Page() {
            @Override
            public int getTotalPages() {
                return 0;
            }

            @Override
            public long getTotalElements() {
                return 0;
            }

            @Override
            public Page map(Function converter) {
                return null;
            }

            @Override
            public int getNumber() {
                return 10;
            }

            @Override
            public int getSize() {
                return 0;
            }

            @Override
            public int getNumberOfElements() {
                return 0;
            }

            @Override
            public List getContent() {
                return Arrays.asList("1");
            }

            @Override
            public boolean hasContent() {
                return false;
            }

            @Override
            public Sort getSort() {
                return null;
            }

            @Override
            public boolean isFirst() {
                return false;
            }

            @Override
            public boolean isLast() {
                return false;
            }

            @Override
            public boolean hasNext() {
                return false;
            }

            @Override
            public boolean hasPrevious() {
                return false;
            }

            @Override
            public Pageable nextPageable() {
                return null;
            }

            @Override
            public Pageable previousPageable() {
                return null;
            }

            @Override
            public Iterator iterator() {
                return null;
            }
        });

        assertEquals(10,pageDto.getCurrentPage());
        assertEquals(0, pageDto.getTotalPages());
        assertEquals(0,pageDto.getTotalElements());


       assertEquals(Arrays.asList("1"),pageDto.getData());



    }
}